cog-snaplookup/requirements.txt
py-cord==2.4.0
python-dotenv==1.0.0
Pillow==9.4.0
tinydb==4.7.1
requests==2.28.2